requirements.txt
hydra-core==1.3.2
invoke==2.2.0
loguru==0.7.3
Markdown==3.7
matplotlib==3.8.2
omegaconf==2.3.0
pydantic==2.10.5
pytorch_lightning==2.5.0.post0
torch==2.5.1
torchvision==0.20.1
typer==0.15.1
wandb==0.19.3
